Refactor: relocate user/host helpers to userhost.go
