+++ cli/command_snapshot_create.go
@@ -2,10 +2,7 @@ package cli
 
 import (
 	"context"
-	"os"
-	"os/user"
 	"path/filepath"
-	"runtime"
 	"strings"
 	"time"
 
@@ -190,46 +187,6 @@ func getLocalBackupPaths(ctx context.Context, rep *repo.Repository) ([]string, e
 	}
 
 	return result, nil
-}
-
-func getUserName() string {
-	return userName
-}
-
-func getDefaultUserName() string {
-	currentUser, err := user.Current()
-	if err != nil {
-		log.Warningf("Cannot determine current user: %s", err)
-		return "nobody"
-	}
-
-	u := currentUser.Username
-
-	if runtime.GOOS == "windows" {
-		if p := strings.Index(u, "\\"); p >= 0 {
-			// On Windows ignore domain name.
-			u = u[p+1:]
-		}
-	}
-
-	return u
-}
-
-func getHostName() string {
-	return hostName
-}
-
-func getDefaultHostName() string {
-	hostname, err := os.Hostname()
-	if err != nil {
-		log.Warningf("Unable to determine hostname: %s", err)
-		return "nohost"
-	}
-
-	// Normalize hostname.
-	hostname = strings.ToLower(strings.Split(hostname, ".")[0])
-
-	return hostname
 }
 
 func init() {
